feat: reorganise components
