+++ app/components/UI/Predict/views/PredictHome/components/PredictCategoriesSection/index.ts
