+++ app/components/UI/Predict/views/PredictHome/components/PredictLiveNowSection/index.ts
