+++ app/components/UI/Predict/views/PredictHome/components/PredictPopularTodaySection/index.ts
