+++ app/components/UI/Predict/views/PredictHome/components/PredictPortfolioModule/index.ts
