+++ app/components/UI/Predict/views/PredictHome/components/PredictTrendingSection/index.ts
